Add files via upload
This code synchronizes a separately recorded signal and video of a  real-world process, recorded simultaneously but with different start and end times, sampling rates, and time scales. The video is slower and the signal data is longer, with the process beginning shortly after the recording starts. The code identifies the process starting in both signal and video,  aligning them as a reference point. It adjusts the video to a real-world time scale and trims signal data not included in the video duration.  For visualization, if signal data points outnumber video frames, it selects the nearest data points; if fewer, it maps one data point to multiple frames. Users can select specific segments from the video or signal for synchronization. The output video's frame rate, quality, and format are customizable.